+++ pkg/authenticator/authenticator_test.go
@@ -88,7 +88,7 @@ func TestAuthenticator_Authenticate(t *testing.T) {
 				// Set the expiration date to now, and try to authenticate again
 				// This will cause the authenticator to try to refresh the cert
 				authn.PublicCert.NotAfter = time.Now()
-				err = authn.Authenticate(context.Background())
+				err = authn.AuthenticateWithContext(context.Background())
 				assert.NoError(t, err)
 
 				// Check that the cert was renewed
@@ -162,7 +162,7 @@ func TestAuthenticator_Authenticate(t *testing.T) {
 			log.ErrorLogger.SetOutput(&logTxt)
 
 			// Call the main method of the authenticator. This is where most of the internal implementation happens
-			err = authn.Authenticate(context.Background())
+			err = authn.AuthenticateWithContext(context.Background())
 
 			// ASSERT
 			tc.assert(t, authn, err, loginCsr, loginCsrErr, logTxt.String())